sections added, visual and api test updated, feedback applied
