+++ src/config.ts
@@ -1,0 +1,1 @@
+export const baseUrl = "https://tegb-backend-877a0b063d29.herokuapp.com";
